requirements.txt
PyYAML==6.0
pandas==1.3.4
torch==2.2.2
rdflib=6.1.1
owlready2==0.45
ollama==0.4.7